feat: 음성녹음, 발음듣기, 재녹음 설정
